new reader tracker for the haplotig assembly and quiver process: an end-to-end test passed.
