+++ src/py_scripts/fc_rr_hctg_track.py
@@ -1,4 +1,4 @@
-from falcon_unzip.select_reads_from_bam import main
+from falcon_unzip.rr_hctg_track import main
 import sys
 if __name__ == "__main__":
     main(sys.argv)
